refactor(#129): lift the roll-stream primitive into roll-crypto
Moves the generic HKDF roll-stream primitive out of item-gen into
roll-crypto so a world-map-encounter derivation can build its own
domain-separated stream without depending on item-gen.
