+++ libs/game/item-gen/src/index.ts
@@ -3,6 +3,7 @@ export { buildSaltStream } from './build-salt-stream';
 export { getTables } from './get-tables';
 export { rollAffixesFromStream } from './roll-affixes-from-stream';
 export { rollItemFromStream } from './roll-item-from-stream';
+export type { RollStream, WeightedEntry } from '@vers/roll-crypto';
 
 export type {
   AffixConstraints,
@@ -18,7 +19,5 @@ export type {
   RolledAffixes,
   RolledItem,
   RollPosition,
-  RollStream,
   SlotContext,
-  WeightedEntry,
 } from './types';
