+++ libs/game/roll-crypto/package.json
@@ -11,7 +11,8 @@
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "dependencies": {
-    "@noble/hashes": "catalog:"
+    "@noble/hashes": "catalog:",
+    "tiny-invariant": "catalog:"
   },
   "devDependencies": {
     "@types/bun": "catalog:",
