+++ libs/game/roll-crypto/src/build-roll-stream.ts
