+++ libs/game/roll-crypto/src/index.ts
@@ -1,4 +1,5 @@
 export { buildRollDigest } from './build-roll-digest';
+export { buildRollStream } from './build-roll-stream';
 export { deriveAvatarKey } from './derive-avatar-key';
 export type { DeriveAvatarKeyInput } from './derive-avatar-key';
-export type { Population } from './types';
+export type { Population, RollStream, WeightedEntry } from './types';
